NOT-WORKING: Initial commit for character URLs
WARNING: This code is experimental and doesn’t work! It’s part of
development for a branch that isn’t functional yet!

I’m keeping a master list of the guids in a data file that I cut and
paste from. I’m not done importing data in data.js yet and I still have
issues with accessing the JSON object in brain.js.